requirements.txt
# Core dependencies
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3

# Airtable integration
pyairtable==2.3.3

# API utilities
flask-cors==4.0.1
flask-limiter==3.8.0
gunicorn==22.0.0

# Cloud storage
boto3==1.34.0

# Validation and serialization
marshmallow==3.21.3
jsonschema==4.23.0

# Logging and monitoring
python-json-logger==2.0.7
sentry-sdk[flask]==1.40.0

# Async processing
celery==5.4.0
redis==5.0.7

# Testing
pytest==8.3.2
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-flask==1.3.0
pytest-html==4.1.1

# Development tools
black==24.4.2
flake8==7.1.0
isort==5.13.2
mypy==1.11.1
types-requests==2.32.0.20240622

# Documentation
sphinx==7.4.7
sphinx-rtd-theme==2.0.0
flask-swagger-ui==4.11.1

# Utilities
python-dateutil==2.9.0
urllib3==2.2.2
